make schematic creation via POST request
